BasicChatbot/requirements.txt
langchain
langchain_community 
langgraph 
langchain_groq
python-dotenv